Edited solute_parameters json and nanofiltration
