+++ src/nanofiltration/nanofiltration_zo.py
@@ -1,0 +1,155 @@
+# imports
+from pyomo.environ import (
+    ConcreteModel,
+    value,
+    TransformationFactory,
+    units as pyunits,
+    assert_optimal_termination,
+    Block,
+    Constraint,
+    Objective,
+    units
+)
+import json
+import math
+import sys
+from pyomo.network import Arc
+from idaes.core import FlowsheetBlock
+from idaes.core.solvers import get_solver
+from idaes.models.unit_models.translator import Translator
+from idaes.core.util.initialization import propagate_state
+from idaes.models.unit_models import Product, Feed
+from idaes.core import UnitModelCostingBlock
+import idaes.core.util.scaling as iscale
+from pyomo.util.check_units import assert_units_consistent
+from watertap.unit_models.nanofiltration_ZO import NanofiltrationZO
+from watertap.unit_models.pressure_changer import Pump
+from watertap.core.wt_database import Database
+import watertap.property_models.multicomp_aq_sol_prop_pack as props
+from idaes.core.util.scaling import (
+    calculate_scaling_factors,
+    constraint_scaling_transform,
+    unscaled_variables_generator,
+    unscaled_constraints_generator,
+    badly_scaled_var_generator,
+)
+from watertap.core.util.initialization import check_dof
+
+sys.path.append('/Users/nicktiwari/Documents/prommis/src/')
+# print(sys.path)
+
+def nanofiltration(m, Q_in = 0.014877, solute_file = '../../solute_parameters.json'):
+    # Read data from 'solute_parameters.json'
+    with open(solute_file) as f:
+        solute_data = json.load(f)
+
+    # solute list
+    solute_list = list(solute_data.keys())
+    mw_data = {key: solute_data[key]['mw'] for key in solute_list}
+    charge = {key: solute_data[key]['charge'] for key in solute_list}
+
+    m.fs.properties = props.MCASParameterBlock(solute_list=solute_list,
+                                               mw_data=mw_data,
+                                               charge=charge,
+                                               density_calculation=props.DensityCalculation.seawater)
+
+    # create units
+    m.fs.feed_nf = Feed(property_package=m.fs.properties)
+    m.fs.nf = NanofiltrationZO(property_package=m.fs.properties)
+    m.fs.P1 = Pump(property_package=m.fs.properties)
+
+    # connections
+    m.fs.feed_to_p1 = Arc(source=m.fs.feed_nf.outlet, destination=m.fs.P1.inlet)
+    m.fs.p1_to_nf = Arc(source=m.fs.P1.outlet, destination=m.fs.nf.inlet)
+
+    TransformationFactory("network.expand_arcs").apply_to(m)
+
+    # specify flowsheet
+    m.fs.feed_nf.properties[0].pressure.fix(101325)  # feed pressure [Pa]
+    m.fs.feed_nf.properties[0].temperature.fix(273.15 + 25)  # feed temperature [K]
+
+    # properties (cannot be fixed for initialization routines, must calculate the state variables)
+    var_args = {("conc_mass_phase_comp", ("Liq", key)): \
+                solute_data[key]['mass_conc_mg_L']/1000 for key in solute_list} # Convert mg/L to kg/m3
+
+    var_args[("flow_vol_phase", ("Liq"))] = Q_in
+    m.fs.feed_nf.properties[0].total_dissolved_solids
+    m.fs.feed_nf.properties.calculate_state(
+        var_args = var_args,  # feed mass fractions [-]
+        hold_state=True,  # fixes the calculated component mass flow rates
+    )
+    m.fs.feed_nf.properties[0].assert_electroneutrality(defined_state=True,
+                                                     adjust_by_ion='Cl')
+    m.fs.P1.efficiency_pump.fix(0.80)  # pump efficiency [-]
+    m.fs.P1.outlet.pressure[0].fix(10e5)
+
+    # fully specify system
+    m.fs.nf.properties_permeate[0].pressure.fix(101325)
+    m.fs.nf.recovery_vol_phase.fix(0.5)
+
+    for key in solute_list:
+        if key != "Cl":
+            m.fs.nf.rejection_phase_comp[0, "Liq", key].fix(solute_data[key]['rejection_phase_comp'])
+
+    m.fs.nf.rejection_phase_comp[0, "Liq", "Cl"] = 0.15  # guess, but electroneutrality enforced below
+    charge_comp = {key: solute_data[key]['charge'] for key in solute_list}
+
+    m.fs.nf.eq_electroneutrality = Constraint(
+        expr=0
+        == sum(
+            charge_comp[j]
+            * m.fs.nf.properties_permeate[0].conc_mol_phase_comp["Liq", j]
+            for j in charge_comp
+        )
+    )
+    constraint_scaling_transform(m.fs.nf.eq_electroneutrality, 1)
+
+    def inverse_order_of_magnitude(number):
+        if number == 0:
+            return "undefined"  # The order of magnitude for zero is undefined
+        magnitude = math.floor(math.log10(abs(number)))
+        return 10**(-magnitude)
+
+    # scaling
+    m.fs.properties.set_default_scaling(
+        "flow_mass_phase_comp", 1e-2, index=("Liq", "H2O")
+    )
+
+    # Set the scaling to be the inverse of the order of magnitude of the mass concentration
+    for key in solute_list:
+        m.fs.properties.set_default_scaling(
+            "flow_mass_phase_comp", inverse_order_of_magnitude(solute_data[key]['mass_conc_mg_L']/1000), index=("Liq", key)
+         )
+
+    m.fs.nf.feed_side.properties_in[0].total_dissolved_solids
+    m.fs.nf.feed_side.properties_out[0].total_dissolved_solids
+
+    m.fs.nf.properties_permeate[0].total_dissolved_solids
+    iscale.set_scaling_factor(m.fs.P1.control_volume.work, 1e-3)
+
+    iscale.calculate_scaling_factors(m)
+
+    # initialize
+    m.fs.feed_nf.initialize()
+    propagate_state(m.fs.feed_to_p1)
+    m.fs.P1.initialize()
+    propagate_state(m.fs.p1_to_nf)
+    m.fs.nf.initialize()
+
+    return m
+
+def main():
+    model = ConcreteModel()
+    model.fs = FlowsheetBlock(dynamic=False)
+    nanofiltration(model)
+    solver = get_solver()
+    solver.solve(model)
+
+    model.fs.nf.report()
+    model.fs.nf.properties_permeate[0.0].display()
+
+    print('='*10 + 'Total Dissolved Solids' + '='*10 +f'\n {model.fs.nf.properties_permeate[0].total_dissolved_solids.value}')
+    return model
+
+if __name__ == "__main__":
+    m = main()
